Add action and role management columns to member table, encapsulating member removal and role checking logic
