+++ app/[alias]/(dashboard)/members/_table/columns.tsx
@@ -1,35 +1,19 @@
 'use client';
 import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar';
-import { Button } from '@/components/ui/button';
-import {
-  Dialog,
-  DialogClose,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle
-} from '@/components/ui/dialog';
 import { Skeleton } from '@/components/ui/skeleton';
-import { cn, formatAddress } from '@/lib/utils';
+import { formatAddress } from '@/lib/utils';
 import { MemberT } from '@/services/chain-db/members';
 import {
-  BundlerService,
   CommunityConfig,
-  Config,
-  hasRole as CWCheckRoleAccess,
-  MINTER_ROLE,
-  waitForTxSuccess
+  Config
 } from '@citizenwallet/sdk';
 import { ColumnDef } from '@tanstack/react-table';
-import { ethers, JsonRpcProvider, Wallet } from 'ethers';
-import { Check, Copy, Trash, X } from 'lucide-react';
+import { ethers } from 'ethers';
+import { Check, Copy } from 'lucide-react';
 import Link from 'next/link';
-import { useRouter } from 'next/navigation';
-import { useEffect, useState } from 'react';
-import { toast } from 'sonner';
-import { useSession } from 'state/session/action';
-import { unpinAction } from '../[account]/action';
+import { useState } from 'react';
+import ActionColumn from './action-column';
+import HasRoleColumn from './has-role-column';
 
 
 const IDRow = ({ account }: { account: string }) => {
@@ -138,66 +122,7 @@ export const createColumns = (
     },
     {
       header: 'Has mint role access?',
-      cell: function Cell({ row }) {
-        const [hasMintRole, setHasMintRole] = useState<boolean | null>(null);
-        const [isPending, setIsPending] = useState(true);
-        const account = row.original.account;
-        const tokenAddress = communityConfig.primaryToken.address;
-        const primaryRpcUrl = communityConfig.primaryRPCUrl;
-
-        useEffect(() => {
-          const checkMintRole = async () => {
-            try {
-              setIsPending(true);
-              const rpc = new JsonRpcProvider(primaryRpcUrl);
-
-              const hasRole = await CWCheckRoleAccess(
-                tokenAddress,
-                MINTER_ROLE,
-                account,
-                rpc
-              );
-              setHasMintRole(hasRole);
-            } catch (error) {
-              console.error('Error checking mint role:', error);
-              setHasMintRole(false);
-            } finally {
-              setIsPending(false);
-            }
-          };
-
-          checkMintRole();
-        }, [account, primaryRpcUrl, tokenAddress]);
-
-        if (isPending) {
-          return <Skeleton className="h-4 w-24" />;
-        }
-
-        return (
-          <div className="inline-flex items-center w-[150px]">
-            <span
-              className={cn(
-                'rounded-full font-medium flex items-center justify-center text-sm py-1 px-2.5',
-                hasMintRole
-                  ? 'bg-green-100 text-green-800 dark:bg-green-900/30 dark:text-green-400'
-                  : 'bg-red-100 text-red-800 dark:bg-red-900/30 dark:text-red-400'
-              )}
-            >
-              {hasMintRole ? (
-                <>
-                  <Check className="mr-1 h-3 w-3" />
-                  Yes
-                </>
-              ) : (
-                <>
-                  <X className="mr-1 h-3 w-3" />
-                  No
-                </>
-              )}
-            </span>
-          </div>
-        );
-      }
+      cell: ({ row }) => <HasRoleColumn row={row} communityConfig={communityConfig} />
     },
     {
       header: 'Created at',
@@ -231,116 +156,7 @@ export const createColumns = (
     },
     {
       header: 'Actions',
-      cell: function RemoveCell({ row }) {
-        const [isDialogOpen, setIsDialogOpen] = useState(false);
-        const [isPending, setIsPending] = useState(false);
-        const [, sessionActions] = useSession(config);
-        const router = useRouter();
-
-        const { image, account, username } = row.original;
-        const isAnonymous = username?.includes('anonymous');
-
-        const handleOpenDialog = () => {
-          setIsDialogOpen(true);
-        };
-
-        const handleCloseDialog = () => {
-          setIsDialogOpen(false);
-        };
-
-        const onRemoveMember = async () => {
-          try {
-            setIsPending(true);
-
-            if (image) {
-              await unpinAction(image);
-            }
-
-            const community = new CommunityConfig(config);
-            const bundler = new BundlerService(community);
-
-            const privateKey = sessionActions.storage.getKey('session_private_key');
-            if (!privateKey) {
-              toast.error('Please login to remove a member');
-              setIsPending(false);
-              router.push(`/${config.community.alias}/login`);
-              return;
-            }
-            const signerAccountAddress = await sessionActions.getAccountAddress();
-
-            const signer = new Wallet(privateKey as string);
-
-            const txHash = await bundler.burnProfile(
-              signer,
-              signerAccountAddress || '',
-              account
-            );
-
-            await waitForTxSuccess(community, txHash);
-
-            await new Promise(resolve => setTimeout(resolve, 250));
-
-            toast.success('Member removed successfully');
-            router.refresh();
-
-            handleCloseDialog();
-          } catch (error) {
-            if (error instanceof Error) {
-              toast.error(error.message);
-            } else {
-              toast.error('Could not remove member');
-            }
-          } finally {
-            setIsPending(false);
-          }
-        };
-
-        return (
-          <>
-            {!isAnonymous && (
-              <>
-                <div className="flex items-center gap-2">
-                  <Button
-                    variant="outline"
-                    className="border-red-500 text-red-500 hover:bg-red-500 hover:text-white"
-                    disabled={isPending}
-                    onClick={handleOpenDialog}
-                  >
-                    <Trash size={16} />
-                    Remove
-                  </Button>
-                </div>
-
-                <Dialog open={isDialogOpen} onOpenChange={setIsDialogOpen}>
-                  <DialogContent className="sm:max-w-md">
-                    <DialogHeader>
-                      <DialogTitle>Remove Member</DialogTitle>
-                      <DialogDescription>
-                        Are you sure you want to remove this member?
-                      </DialogDescription>
-                    </DialogHeader>
-                    <DialogFooter className="sm:justify-start gap-2">
-                      <Button
-                        disabled={isPending}
-                        onClick={onRemoveMember}
-                        type="button"
-                        variant="destructive"
-                      >
-                        {isPending ? 'Removing...' : 'Remove'}
-                      </Button>
-                      <DialogClose asChild>
-                        <Button type="button" variant="outline" disabled={isPending}>
-                          Cancel
-                        </Button>
-                      </DialogClose>
-                    </DialogFooter>
-                  </DialogContent>
-                </Dialog>
-              </>
-            )}
-          </>
-        );
-      }
+      cell: ({ row }) => <ActionColumn row={row} config={config} />
     }
   ];
